Introduce new mappings
I propose following changes:
folding with - (hyphen) and ? (question mark, recursively)
task checking with xx (double x)
These should be easier to reach as well as being used infrequently. Tab
is a common shortcut for completion and = is used with the same hand as
hjkl and far away. The new mappings should be more convenient